moved result calculation to PHP
Altered the system flow to allow PHP to generate the result, pass the data to JS to display the result, pass the data to PHP (after spin finished) to be processed. Altered the way the items were given; items are taken from the items directory where the images are stored so that if more items are added, no code needs to be changed.